ACCUMULO-4138 Fix the description of -b options
It was wrong for FlushCommand and CompactCommand.  In the process of
refactoring, the following was accomplished:

- Change the description in the OptUtil.startRowOpt to be (NOT) inclusive.
- Use OptUtil.startRowOpt in CompactCommand, MergeCommand and FlushCommand.
- Refactor the DeleteRowCommand to use the OptUtil.startRowOpt method.
- Refactor ScanCommand to override the description to say it is inclusive.

Also updated the javadoc from a potential cut and paste on TableOperations.

A little mixed on the usefulness of the Unit Tests for these.  They may be
brittle, but perhaps they would alert if someone changed the way one command
worked without looking at the others.