ACCUMULO-4148 Native map doesn't increment counter for every cell, potentially losing some updates
The problem here is that InMemoryMap$DefaultMap increments the mutationCount or
kvCount for every key value pair in every Mutation that is passed in.  The
NativeMap, which is used by the InMemoryMap$NativeMapWrapper does not.  This
causes 2 different issues in the NativeMap.

1)  When a single Mutation has duplicate key value pairs, only the last is
recorded, because they all have the same mutationCount and the earlier ones are
hidden.
2 ) When multiple Mutations are passed in at the same time, the mutationCount
or kvCount starts over for each Mutation.  This can also lead to hiding of key
value pairs.

The tests added here expose both the issues as well as do some asserts on
simple Mutations.  A few tweaks were made to expose information to these tests.

1) Made MemKey public, made it's kvCount private and exposed that via a getter so
we can inspect directly instead of parsing the toString.  Required changing
some calls to kvCount to use the getter.

2) Added a final String to the InMemoryMap which is set during construction.  This
allows you to see what kind of SimpleMap was setup in the InMemoryMap.

Closes apache/accumulo#82

Signed-off-by: Josh Elser <elserj@apache.org>