ACCUMULO-1755: Modified TSBW so that all client threads will not block on binMutations
Before this change all client threads that were adding mutations to a BatchWriter would block
when the mutations needed to be written to the tablet servers. This change gives each client
thread their own MutationWriter object using a ThreadLocal. With this change only one client
should block on adding the mutations to the sendThreadPool object, and allows the other client
threads to push mutations onto a new MutationSet.