ACCUMULO-4229 Isolate Locator sync code; better style
Rather than riddle the client classes that cache TabletLocators with synchronization code,
this commit abstracts the synchronization into a new SyncingTabletLocator class.
A SyncingTabletLocator automatically fetches the current TabletLocator
when the cached TabletLocator becomes outdated (when isValid flips false).

TimeoutTabletLocator now uses the SyncingTabletLocator.
In order to do so this commit specializes TimeoutTabletLocator to its usage in two classes.

Variables assumed invariant for SyncingTabletLocator are now final.
This does not affect semantics.

The new design allows for TabletServerBatchWriter to use its local Locator cache
more aggressively than it used to.