ACCUMULO-4321 Only seal jars on release
Skip jar sealing under normal builds by default, so integration tests
don't generate sealing violation errors. Seal jars in the release
profile.

This is a quick fix. The long-term fix is to ensure any integration
and/or performance tests are in distinct packages so they can be run
even with released jars.