ACCUMULO-4157 Bug fix for removing WALs to quickly
Keep track of first time a tserver is seen down and only remove WALs for that server if past configurated threshhold

Trying to keep the changes small to fix the bug.  I'll create another ticket to refactor and cleanup

Includes an end to end test calling the collect method simulating a dead tserver.