ACCUMULO-4348 Deprecate KerberosToken constructor with side effects
`KerberosToken(String, File, boolean)` is deprecated in favor of `KerberosToken(String, File)`.

The boolean flag would log in the requested user with Hadoop's `UserGroupInformation` class. This
changed global state about who the active user was. In a multi-user environment, this potentially
made little sense as other users could overwrite eachother.

This patch includes a convenience constructor that doesn't have any side effects, but has the same
semantics as logging in a user with a keytab.