ACCUMULO-4351 Default to failIfNoSpecifiedTests=false
Set properties for surefire and failsafe plugins so that the
non-existence of specified tests to run do not cause a build failure.
This makes it easier to execute a specific test which exists only in one
module of the multi-module project, but which does not exist in other
modules.

(Also remove redundant property in 'sunny' profile)