ACCUMULO-4373 Fix compat w/ commons-configuration
Revert to commons-configuration 1.6 dependency instead of the newer one.
This preserves the compile-time binding to the MapConfiguration
constructor with a Map parameter for Properties, rather than the newer
constructor with Properties as a parameter. This allows our compiled
code to work on a wider range of commons-configuration versions.

However, this also encourages the use of older, buggy versions of
commons-configuration, and forces the use of a constructor which does
not properly filter Properties to only make use of those with String
keys, even when a newer version of commons-configuration is used.

A newer version of commons-configuration should be used when it can be
reasonably expected that user's environments have the newer version.
(Effectively, this means, when our Hadoop dependency switches to a newer
version of commons-configuration.)