ACCUMULO-4574 Modified TableOperations online to check if table is already online before executing fate transaction.
When the online command is issued, the fate operation will block if a fate transaction has locked the table. If
the table is already online, there is no reason to block and then issue the online operation. This modification
turns the online command into a noop if the table is already online. This change includes an IT
test that uses a compaction with slow iterator to cause the fate transaction to lock a table and
then runs the online command - checking that the operation did not block.