[ACCUMULO-4506] Add a timeout to a replication RPC call
This addresses an issue where a replication task will get stuck for a
substantial amount of time.  Using a timeout for the client will abandon
the task after 2m (default) for a single RPC attempt.  Note that this is
related to the replication.work.attempts property, as a client timeout
will be retried by the same task up to this number of times before being
abandoned entirely.