ACCUMULO-4708 Limit RFile block size to 2GB
Moved the configuation sanity check to ConfigSanityCheck instead of unit tets
Added a >= to the BCFile stream size check
Added unit tests for a key that is too large or a (key, value) pair that is too large. This resulted in a modification to the pom.xml to allow users to configure memory.