ACCUMULO-2907 Remove warning from initialize (#301)
Updated console message in the getRootPassword method to remove unneeded warning about the security setup.

Modified getRootPassword method of Initialize.java to appropriately
display an security warning message when asking for initial password
setup. If the value of default value of instance.security.authentication
property differs from the Site configured value then a warning is
displayed indicating that the security setup may not be applicable.