ACCUMULO-2968: Axe TabletServer.majorCompactorDisabled (#302)
* ACCUMULO-2968: Axe TabletServer.majorCompactorDisabled

Removed references to majorCompatorDisabled variable defined in
accumulo/tserver/TabletServer.java.  This private variable was
defined as 'false' and never modified in source and additionally
no means were provided to alter the value in the code.

Several locations within TabletServer.java where updated. This
included a couple of 'if' statement that would never be executed as well
as some return statement where a constant value of true was
unneccessarily included.

There was one reference each with CompationRunner.java and Tablet.java
which were removed also.

The change of most note is probably the while loop within the run method
in TableServer. Previously the while check read:
   while(!majorCompatorDisabled)....

whereas now it reads:
   while(true)...

which was what it would always equate too. This could be seen to be a
loss of information perhaps. A comment was added to describe what
existed prior to the update.