ACCUMULO-4170 Clarify ClientConfiguration javadocs
Updated the javadoc information with the ClientConfiguration class.
Specifically reworked the default search path information to be
displayed as a list rather than inline, thereby easing readability.

Reworded a few of the sentences and moved the path order up to the class
level rather than the method level.

Updated to revert the default search order information from the class
level back into the loadDefaults() method level.