ACCUMULO-4732 No APIs to configure iterators and locality groups for new table
Added several methods to the NewTableConfiguration class to all for the configuration of locality groups
and the attachement of iterators prior to a tables creation.

Modified/created 5 files:

NewTableConfiguration.java now has a setLocalityGroups, and two variations of an attachIterator method as
well as an additional helper method for checking iterator conflicts.

There is a new class, NewConfigurationTestIT.java to verify the updates made to the API.

An ensureNonOverlappingGroups functionality was extracted from TableOperationsImpl.java and added to
LocalityGroupUtil.java.

ReadWriteIT.java has an additional locality group test as well.