ACCUMULO-4756 Small regex fixes for validating names
Fix a few small regex issues with namespace and table name validation.
This fix allows the default namespace to be specified in the
comma-separated list of namespaces, along with other namespaces.

Also apply minor regex cleanup by removing unnecessary parens, applying
constraints on the number of digits in a port for server:port, and
deduplicating the server regex for the case where it is optional.

Also apply previously uncommitted formatting changes.