ACCUMULO-4788 Make connection pool O(1) instead of O(N) (#385)
The Jira issue mentions two problems inefficiency and a global lock. This commit focuses on the inefficiency issue and makes the reserve and return operations O(1) instead of O(N). The global lock is stil there, but now operations done while the lock are held are much faster. I plan to open up a follow on issue to address the lock.