Fix ShellServerIT test in master
The master branch needed several quick IT test updates to ShellServerIT to pass successfully with the
recent changes to the setshelliter command. These tests did not exist in the 1.8 line and were not updated
with the recent 1.8 merge. They involved the '-t table' option from several setshelliter statments in the
tests.