Fix #461 Handle race condition in AccumuloMonitorAppenderTest
Catch NoSuchElementException and handle it in the
AccumuloMonitorAppenderTest. This can happen if the test sees the state
of the appender while it is being updated from the old monitor location
to the new location.

Updates to the appender occurs in a single thread. Dispatching log
events are synchronized, and thread-safe, so this only ever affected the
test itself.