ACCUMULO-2537 Add @Override Annotations
Added a profile 'use-errorprone' to enforce use of the override annotation on
required methods. Uses google errorprone to handle enforcement. It
ignores the generated code in the project. Findbugs/Spotbugs do not
currently enable enforcement of the annotation, thus the use of
errorprone. Initially setup as a profile to verify it works as expected
and satifies the goal of the annotation enforcement.

Addressed some issues brought up via Github review. Included removal of
some unnecessary tags and the addition of a comment explaining the
profile. Also renamed to 'use-errorprone'.