Update Thrift to latest (0.11.0)
Remove workarounds for bugs fixed in 0.11.0:
* Remove workaround for spammy debug print statement (THRIFT-4062)
* Remove workaround for unused imports in enums (THRIFT-4073)
* Remove RpcWrapper (more details below)

***
RpcWrapper was added in ACCUMULO-2950 to catch server-side
RuntimeExceptions and notify the client with a TApplicationException, as
in Thrift 0.9.0 (See THRIFT-1805). It is no longer needed since a flag
was added to the generator to optionally handle RuntimeExceptions this
way, which we now set in our generate-thrift.sh.