Introduce a property for a key password for the Monitor keystore file
Java keystores can have a keystore password as well as a password on the private key contained inside of that keystore
(when that keystore may contain multiple keys). Usually, but not always, the "key password" is the same as the
keystore password. This commit introduces a new property to allow this configuration.