Roll back maven-invoker-plugin 3.1.0->3.0.1
A bug exists in maven-invoker-plugin 3.1.0 which causes the maven-plugin
module to fail to build, due to missing SNAPSHOT dependencies copied
into the PluginIT's local repository from sibling modules. This is
because transitive dependencies fail to be copied, as they were in 3.0.1
of the plugin.

See https://issues.apache.org/jira/browse/MINVOKER-243