Fix Travis conflict with rat plugin
The travis_wait command in Travis CI causes maven-rat-plugin check
failures, so this modifies the build command to only run that for the
spotbugs/findsecbugs task that disables rat check.

In addition to the Travis console timeout, which was increased to 20
minutes, spotbugs has its own timeout, which must be increased for
findsecbugs, which runs slowly in Travis CI's infrastructure. That is
now increased to 20 minutes from the default of 10 as well.