fixes #587 Check deleting state when deciding on tablets
The following two changes were made :

 * In tableOps.DeleteTable.call() the table state is transitioned to
   DELETING.  This commit ensures the table state cache is updated.
   These are the changes in TableManager.java.
 * The main loop in TabletGroupWatcher now checks if tables are DELETED
   or DELETING.  It used to only check for DELETED (which is null).

These changes rely on the fact that tableOps.CleanUp.isReady() calls
master.hasCycled().  This ensures the current loop that TabletGroupWatcher
is running finishes before CleanUp starts deleting metadata entries. With
these changes, the next TabletGroupWatcher loop should see the table state
changes and not add any junk to the metadata table.

Also, slightly unrelated to this bug fix I noticed that the transition table
state method in TabletGroupWatcher was not idempotent. Specifically if a FATE
delete table op attempted to transition a table to DELETING that was already
DELETING it would have thrown an error.  Now transitioning DELETING to
DELETING should be a noop.