Fixes #587 Wait for tablets assigned to dead servers before deleting table
The tableOps.CleanUp.isReady() operation was not waiting on tablets that were
assigned to dead tablet servers.  The master will attempt to write logs and
delete the location for these tablets.  Therefore its best to wait for the
Master to make its metadata update before proceeding to clean up a deleted
table.