fixes #587 In table delete wait for tablets assigned to dead servers (#727)
The tableOps.CleanUp.isReady() operation was not waiting on tablets that were
assigned to dead tablet servers.  The master will attempt to write logs and
delete the location for these tablets.  Therefore its best to wait for the
Master to make its metadata update before proceeding to clean up a deleted
table's metadata.  Not waiting could result in junk in the metadata table.

Also, slightly unrelated to this bug fix I noticed that the transition table
state method in TabletGroupWatcher was not idempotent. Specifically if a FATE
delete table op attempted to transition a table to DELETING that was already
DELETING it would have thrown an error.  Now transitioning DELETING to
DELETING should be a noop.