fixes #785 avoid using system classloader in client rfile code (#786)
* fixes #785 avoid using system classloader in client rfile code

Properly making this change cascaded into wider changes.  Trying to avoid the
system classloader led to the discovery of code in BCFile that created a
crypto service using static config and the system classloader.  I realized
this code needed to go which resulted in making the crypto service mandatory
when opening an RFile.  This led to pushing the  ServerContext further into
code, which is a good thing.

The following actions were taken :

 * Made CryptoService a required argument for opening a file in FileOperations
 * MAde CryptoServiceFactory require the caller to specify if accumulo
   classloader should be used.
 * Added CryptoServiceFactory.newDefaultInstance() method and called this in
   places like test code that just need the default crypto service.
 * Pushed ServerContext further into code.  One nice thing about this is that
   things available from ServerContext no longer needed to passed.