Refactor FATE AdminUtil spliting getStatus method.
Refactor AdminUtil to provide refactor the existing fate status methods into two - one that provides the existing behaviour and a second
that does not need to process lock information when that information is not required. This refactoring is a step to providing FATE metrics
where a count of existing FATES does not need the lock information.
* Adds unit test
* Includes code analysis clean-up.