Fix race condition when getting WALs for dead tserver (#866)
When a tablet server dies the master gets its WALs from ZK.  In ZK there
is a list of WALs per tserver. Each WAL in ZK has state that is either
OPEN, CLOSED, or UNREFERENCED.  The master needs a list of OPEN and
CLOSED logs for dead tservers.  While the master is trying to obtain
this list its possible that the Accumulo GC may delete an UNREFERENCED
WAL.  If this happened then the code before this commit would return an
empty list of WALs. This could result in OPEN and CLOSED logs being
ignored for recovery which could result in data loss.  This patch fixes
the race condition.

This bug was observed while looking into an exception I noticed while
writing test for #860.  In the IT I was frequently calling another
WalStateManager function and noticed NoNodeExceptions.  As a result I
examined how the entire class handled NoNode race conditions and found
this bug.  I noticed some other possible NoNode race condition, but do
not think this these would occur with the current way the code is called.