During tablet recovery filter logs out that do not define tablet.
In a situation where a tablet had lots of write ahead logs, log
recovery consumed a lot of memory.  The reason for the memory
consumption was that all recovery logs are opened at once.  This
means the index for each file and a few key/values from each file
are read into memory.

This patch modifies the code to first open logs one at time to see if
they define the tablet. Then only the logs that define the tablet are
opened at the same time.  This has the potential to use less memory
when a subset of the logs define a tablet.