Improve log4j lookup
Previously, I used cdh5.x hadoop version is 2.x, and then I found the log4j jar when I executed accumulo init. But when I upgraded to cdh6 hadoop version 3.0, log4j class not found and it was debugged. The shell looks for log4j-core-*.jar and is not log4j-1.2.17.jar