Avoid exposing internal types in new VolumeChooser SPIs
For 2.0.0 new method were added to VolumeChooser that expose internal
types like ServerContext.  These types are very unstable and will
likely change alot in the future as we seek to remove static state
in Accumulo.  This PR introduces a simple SPI interface that exposes
some of the functionality of ServerContext in a more stable way.