Refactor FATE AdminUtil spliting getStatus method. (#891)
Refactor AdminUtil to provide refactor the existing fate status methods into two - one that provides the existing behaviour and a second
that does not need to process lock information when that information is not required. This refactoring is a step to providing FATE metrics
where a count of existing FATES does not need the lock information.
* Adds unit test
* Includes code analysis clean-up.

* AdminUtil_refactor - updated with pull request commnets

This refatoring is to provide a new lighter weigth method to get FATE status
for metrics. This commit tries to address pull request commnets for review.
There may be additional clean-up before final commit, but wanted to make sure
this was on track.

* Updated to address latest comments on pull request.

Modified the utility and the tests to address commnets by keith-turner.

* Corrected log statment that was printing timing info before actual call.

This address comment by Christopher that caught a log statement that
was inadverntally moved as part of clean-up.

* AdminUtil_refactor - renamed test to clarify the types of IT tests run.

* corrected javadoc formatting