Ignore unused variables in try-with-resources
Reduce compiler warning spam from the use of AutoCloseable trace objects
created in try-with-resources blocks. This could allow some bugs to be
overlooked of the type where the wrong variable is referenced inside the
block or the resource is created, but not needed for the block.
Unfortunately, it's not possible to ignore them by specific types.
Otherwise, I'd just ignore the TraceScope ones.

It may be useful to, instead of ignoring these warnings for TraceScope
objects, add annotated data to the scope, so that way it's both
referenced, and has additional meaningful data in the resulting trace
span.