Improve sig/checksum wording on rc template
Update the release candidate build script to improve the phrasing in the
release candidate email regarding cryptographic signatures and
references to checksums for Maven/Nexus tooling.

The intent is to make it more clear what purpose the sha1/md5 files
serve, and to more prominently highlight the cryptographic pieces that
are significantly more relevant to release prep and voting, but without
being excessively verbose.