Create AbstractServer type for Accumulo servers
Create base class to help standardize the way Accumulo servers are
initialized from configuration files, launch, run, and clean up
resources during shutdown.

This establishes the base class as a concept, but over time, more
implementation that servers have in common can be put here, to make
server-side code more maintainable, and so the code in the individual
servers can be limited to the bits that make them different.