fixes #1128 cleaned up upgrade code
Removed a lot of upgrade code for older versions that should never run.
Refactored most of the upgrade code out of the master.  One goal of the
refactoring was to create a structure for upgrades that ensures the
upgrade code for each version is in a separate file.  The existing code
for upgrading zookeeper was single method with stuff for lots of
different versions.  This made understaning what upgrade steps went with
which versions very difficult.