Remove custom property interpolations
Remove unnecessary custom property interpolations in code. Instead, rely
on property interpolation features of commons-configuration2, as needed,
to set properties based on environment.

A side effect of this change, as well as the fact that
GENERAL_DYNAMIC_CLASSPATHS is deprecated, is that the default value for
that property is now blank (and the empty lib/ext directory is no longer
created in the assembly tarball).

In future, commons-configuration2 could also be used to interpolate
values of configuration properties stored in ZooKeeper, but that is out
of scope of this issue, and can be added later, if there is interest.

This change further decouples the reliance on specialized environment,
and reduces assumptions made in our code, making it more
environment-agnostic.