Ensures correct use of ZooKeeper getAcl
This fixes a bug in Accumulo where we were using an unauthenticated
ZooKeeper session to read ACLs on a node that we had previously written.

ZooKeeper 3.4.14 fixed a bug in their code that allowed unauthenticated
connections to read ACLs on paths for which they did not have READ
permission.

Using Accumulo with that version of ZooKeeper exposed the bug in our
code that passed the incorrect ZooKeeper session when retrieving ACLs.