Improve use of Hadoop's DistributedCache (#1188)
Update DistributedCacheHelper to both add a cached file to a MapReduce
job and to retrieve an InputStream from the DistributedCache (or HDFS,
if the item was not successfully cached).

Ensure uniqueness in the local cached file names and that the files are
read from the correct filesystem.

Add some tests to TokenFileIT

Specify the charset explicitly when java.util.Scanner is used (most
occurrences were related to code using the DistributedCache, but also
fixed two trivial occurrences in shell-related code).