Revert "De-dupe some code (#1211)"
This reverts commit 417de4e6da1f03c18e3717f2c359828f500197b2.

This is reverted because it breaks two tests in ZooKeeperInstanceTest,
one of which can be updated to test the equivalent scenario in ZooUtil,
but the other test breakage reveals that this change fails to properly
handle the the lookup when the ZooKeeperInstance is constructed with the
instanceId (instanceName is null) and the instanceId does not correspond
to a valid location in ZooKeeper. This situation should result in a
RuntimeException, but it does not.