Fix #1212 re-apply #1211 preserving prior behavior
Revert "Revert "De-dupe some code (#1211)""

This reverts commit ca5b62d81b21a35bb260f1c5d3257811eeb9ae6e.
(This reverts the revert, and then adds fixes)

Preserve behavior prior to #1211 that verifies (using ZooCache) that the
instanceId actually points to a valid instanceId path in ZooKeeper.
This verification is done regardless of whether instanceId is resolved
using the instanceName, or if it is set explicitly in client
configuration (in the case of ZooKeeperInstance). It also verifies on
each call to getInstance() (just in case the instance is deleted from
ZooCache unexpectedly, things won't continue operating as though
everything is fine).