fix #1247 store fate txid in ~blip marker value
Before this change the bulk import code was serializing and logging fate
transaction ids differently than other parts of the code.  When adding
FATE transaction ids to blip markers, I decided it would be best if they
were consistently persisted (in metadata table ) and logged. That
decision made this change much larger than simply adding something to
the blip value, but I feel its worth it.

The following are the main changes :

 * Introduce new class FateTxId with methods for consistenly formatting
   fate transaction ids.
 * Make bulk fate logging use FateTxId
 * Update ~blip and loaded family to use FateTxId.  Made loaded family
   backwards compat.
 * Update existing fate logging code to use FateTxId.  Instead of
   String.format

Follow on work to update all Repos to use FateTxId would be nice.  Chose
not do to that in the change because it would make it harder to review.