correctly handle first and last tablets in LinkingIterator #1260
The LinkingIterator checks the structure of the AccumuloMetadata table
as it is scans.   It was not properly handling a split of the first
tablet.  It was also not ensuring the last tablet for a table was seen.
This patch fixes those two issues.  These issues were discovered while
researching #1260, however I am not sure if these issues could have
caused #1260.