Refactor how constraint violations are handled
Refactor out the usage of exceptions as a means of returning information
about constraint violations.

Tablet.prepareMutationsForCommit threw an exception if constraint
violations were found in any mutations and the tablet server was
available for write. Information about the violations were pulled from
the exception for logging and any additional commit preparation. Using
the exception as a means of control flow is an anti-pattern.

Wrap the CommitSession in TabletMutationPrepAttempt that contains all
identified violations and non-violators and explicitly check and handle
any violations in the main body of pertinent functions in TabletServer.

TConstraintViolationException.java is no longer used. Delete it.

Fixes #824