Improved logging of busiest tablets.
Now the deltas of tablets ingest and query counts are used for computing
the busiest N tablets. The previous code used the absolute counts.
Restructured code that so that it is unit testable and found a few bugs
while writing the unit test.  Found the prioQ was not properly tracking
the top N tablets.