Update javadoc:aggregate config to match public API
No longer exclude subpackages from javadoc generation for the public API
on the website. The API was simplified in 2.0 to include all
subpackages, and any undesired subpackages were moved to a different
package hierarchy. So, there's no reason to exclude them in the javadoc
configuration. With this, the javadoc:aggregate configuration matches
the declared public API definition.