Use concurrent map instead of cache for bulkImported map
While analyzing some tablet code I noticed a Guava cahce was being used
like a map.  When I first saw the code I was concerned the cache may
evict entries (which would be a bug).  However the cache was created
with default settings, which according to the javadoc does not
automatically evict.  This commit replaces the cache with a concurrent
map so no one has to do this analysis again.