Give ThriftTransportPool higher concurrent granularity
Instead of locking the entire cache when performing an operation for a
single ThriftTransportKey, make it possible to lock access to the cached
connections for the key in question without blocking operations for
other keys.

In addition, refactor sections of the class to improve the clarity of
intent for operations that obtain locks individually for each transport
key, and add documentation to this effect.

Work for #1122