Always choose volume for new tablet files.
This commit changes Accumulo to always call the volume chooser every time a
tablet creates a new file.  It also changes the interpretation of the srv:dir
column in the metadata table.  This column used to contain a URI to a
directory on a specific volume that was used for all new tablet files. Now the
srv:dir column only contains a directory name.  This directory name will be
used for new tablet files across all volumes.

This change necessitated to ~del markers in the metadata table used for
garbage collection.  When a table is cloned or tablets are merged out of
existance it can result in ~del markers for tablet dirs being placed in
the metadata table.  These ~del markers used to reference a specific volume.
With this change, the ~del marker now use a special URI  of the form

  accumulo://allVolumes/accumulo/tables/<tableId>/<dir name>

When the Accumulo GC sees this, it will delete the dir on all configured
volumes when its no longer used.

This change superceded #642.  These changes are possible because of the
changes made in #936.